fix: Link redirects for original links. (#63)
If a user goes to one of the doc links for the old structure, it should redirect to the new one. For example:

https://docs.dgraph.io/deploy/#bulk-loader should redirect to
http://docs.dgraph.io/docs//deploy/fast-data-loading/#bulk-loader

The previous version of this code doesn't work when the docs are served as a subpage (i.e., https://dgraph.io/docs or localhost/docs instead of simply localhost).